+++ src/apify_client/clients/resource_clients/request_queue.py
@@ -1,7 +1,9 @@
-from typing import Any, Dict, Optional
+from typing import Any, Dict, List, Optional
+
+from apify_shared.utils import filter_out_none_values_recursively, ignore_docs, parse_date_fields
 
 from ..._errors import ApifyApiError
-from ..._utils import _catch_not_found_or_throw, _filter_out_none_values_recursively, _parse_date_fields, _pluck_data, ignore_docs
+from ..._utils import _catch_not_found_or_throw, _pluck_data
 from ..base import ResourceClient, ResourceClientAsync
 
 
@@ -44,7 +46,7 @@ class RequestQueueClient(ResourceClient):
             'name': name,
         }
 
-        return self._update(_filter_out_none_values_recursively(updated_fields))
+        return self._update(filter_out_none_values_recursively(updated_fields))
 
     def delete(self) -> None:
         """Delete the request queue.
@@ -72,7 +74,30 @@ class RequestQueueClient(ResourceClient):
             params=request_params,
         )
 
-        return _parse_date_fields(_pluck_data(response.json()))
+        return parse_date_fields(_pluck_data(response.json()))
+
+    def list_and_lock_head(self, *, lock_secs: int, limit: Optional[int] = None) -> Dict:
+        """Retrieve a given number of unlocked requests from the beginning of the queue and lock them for a given time.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/queue-head-with-locks/get-head-and-lock
+
+        Args:
+            lock_secs (int): How long the requests will be locked for, in seconds
+            limit (int, optional): How many requests to retrieve
+
+
+        Returns:
+            dict: The desired number of locked requests from the beginning of the queue.
+        """
+        request_params = self._params(lockSecs=lock_secs, limit=limit, clientKey=self.client_key)
+
+        response = self.http_client.call(
+            url=self._url('head/lock'),
+            method='POST',
+            params=request_params,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))
 
     def add_request(self, request: Dict, *, forefront: Optional[bool] = None) -> Dict:
         """Add a request to the queue.
@@ -98,7 +123,7 @@ class RequestQueueClient(ResourceClient):
             params=request_params,
         )
 
-        return _parse_date_fields(_pluck_data(response.json()))
+        return parse_date_fields(_pluck_data(response.json()))
 
     def get_request(self, request_id: str) -> Optional[Dict]:
         """Retrieve a request from the queue.
@@ -117,7 +142,7 @@ class RequestQueueClient(ResourceClient):
                 method='GET',
                 params=self._params(),
             )
-            return _parse_date_fields(_pluck_data(response.json()))
+            return parse_date_fields(_pluck_data(response.json()))
 
         except ApifyApiError as exc:
             _catch_not_found_or_throw(exc)
@@ -150,7 +175,7 @@ class RequestQueueClient(ResourceClient):
             params=request_params,
         )
 
-        return _parse_date_fields(_pluck_data(response.json()))
+        return parse_date_fields(_pluck_data(response.json()))
 
     def delete_request(self, request_id: str) -> None:
         """Delete a request from the queue.
@@ -169,6 +194,112 @@ class RequestQueueClient(ResourceClient):
             method='DELETE',
             params=request_params,
         )
+
+    def prolong_request_lock(self, request_id: str, *, forefront: Optional[bool] = None, lock_secs: int) -> Dict:
+        """Prolong the lock on a request.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/request-lock/prolong-request-lock
+
+        Args:
+            request_id (str): ID of the request to prolong the lock
+            forefront (bool, optional): Whether to put the request in the beginning or the end of the queue after lock expires
+            lock_secs (int): By how much to prolong the lock, in seconds
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+            forefront=forefront,
+            lockSecs=lock_secs,
+        )
+
+        response = self.http_client.call(
+            url=self._url(f'requests/{request_id}/lock'),
+            method='PUT',
+            params=request_params,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))
+
+    def delete_request_lock(self, request_id: str, *, forefront: Optional[bool] = None) -> None:
+        """Delete the lock on a request.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/request-lock/delete-request-lock
+
+        Args:
+            request_id (str): ID of the request to delete the lock
+            forefront (bool, optional): Whether to put the request in the beginning or the end of the queue after the lock is deleted
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+            forefront=forefront,
+        )
+
+        self.http_client.call(
+            url=self._url(f'requests/{request_id}/lock'),
+            method='DELETE',
+            params=request_params,
+        )
+
+    def batch_add_requests(self, requests: List[Dict[str, Any]], *, forefront: Optional[bool] = None) -> Dict:
+        """Add requests to the queue.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/batch-request-operations/add-requests
+
+        Args:
+            requests (List[Dict[str, Any]]): List of the requests to add
+            forefront (bool, optional): Whether to add the requests to the head or the end of the queue
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+            forefront=forefront,
+        )
+
+        response = self.http_client.call(
+            url=self._url('requests/batch'),
+            method='POST',
+            params=request_params,
+            json=requests,
+        )
+        return parse_date_fields(_pluck_data(response.json()))
+
+    def batch_delete_requests(self, requests: List[Dict[str, Any]]) -> Dict:
+        """Delete given requests from the queue.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/batch-request-operations/delete-requests
+
+        Args:
+            requests (List[Dict[str, Any]]): List of the requests to delete
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+        )
+
+        response = self.http_client.call(
+            url=self._url('requests/batch'),
+            method='DELETE',
+            params=request_params,
+            json=requests,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))
+
+    def list_requests(self, *, limit: Optional[int] = None, exclusive_start_id: Optional[str] = None) -> Dict:
+        """List requests in the queue.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/request-collection/list-requests
+
+        Args:
+            limit (int, optional): How many requests to retrieve
+            exclusive_start_id (str, optional): All requests up to this one (including) are skipped from the result
+        """
+        request_params = self._params(limit=limit, exclusive_start_id=exclusive_start_id, clientKey=self.client_key)
+
+        response = self.http_client.call(
+            url=self._url('requests'),
+            method='GET',
+            params=request_params,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))
 
 
 class RequestQueueClientAsync(ResourceClientAsync):
@@ -210,7 +341,7 @@ class RequestQueueClientAsync(ResourceClientAsync):
             'name': name,
         }
 
-        return await self._update(_filter_out_none_values_recursively(updated_fields))
+        return await self._update(filter_out_none_values_recursively(updated_fields))
 
     async def delete(self) -> None:
         """Delete the request queue.
@@ -238,7 +369,30 @@ class RequestQueueClientAsync(ResourceClientAsync):
             params=request_params,
         )
 
-        return _parse_date_fields(_pluck_data(response.json()))
+        return parse_date_fields(_pluck_data(response.json()))
+
+    async def list_and_lock_head(self, *, lock_secs: int, limit: Optional[int] = None) -> Dict:
+        """Retrieve a given number of unlocked requests from the beginning of the queue and lock them for a given time.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/queue-head-with-locks/get-head-and-lock
+
+        Args:
+            lock_secs (int): How long the requests will be locked for, in seconds
+            limit (int, optional): How many requests to retrieve
+
+
+        Returns:
+            dict: The desired number of locked requests from the beginning of the queue.
+        """
+        request_params = self._params(lockSecs=lock_secs, limit=limit, clientKey=self.client_key)
+
+        response = await self.http_client.call(
+            url=self._url('head/lock'),
+            method='POST',
+            params=request_params,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))
 
     async def add_request(self, request: Dict, *, forefront: Optional[bool] = None) -> Dict:
         """Add a request to the queue.
@@ -264,7 +418,7 @@ class RequestQueueClientAsync(ResourceClientAsync):
             params=request_params,
         )
 
-        return _parse_date_fields(_pluck_data(response.json()))
+        return parse_date_fields(_pluck_data(response.json()))
 
     async def get_request(self, request_id: str) -> Optional[Dict]:
         """Retrieve a request from the queue.
@@ -283,7 +437,7 @@ class RequestQueueClientAsync(ResourceClientAsync):
                 method='GET',
                 params=self._params(),
             )
-            return _parse_date_fields(_pluck_data(response.json()))
+            return parse_date_fields(_pluck_data(response.json()))
 
         except ApifyApiError as exc:
             _catch_not_found_or_throw(exc)
@@ -316,7 +470,7 @@ class RequestQueueClientAsync(ResourceClientAsync):
             params=request_params,
         )
 
-        return _parse_date_fields(_pluck_data(response.json()))
+        return parse_date_fields(_pluck_data(response.json()))
 
     async def delete_request(self, request_id: str) -> None:
         """Delete a request from the queue.
@@ -335,3 +489,108 @@ class RequestQueueClientAsync(ResourceClientAsync):
             method='DELETE',
             params=request_params,
         )
+
+    async def prolong_request_lock(self, request_id: str, *, forefront: Optional[bool] = None, lock_secs: int) -> Dict:
+        """Prolong the lock on a request.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/request-lock/prolong-request-lock
+
+        Args:
+            request_id (str): ID of the request to prolong the lock
+            forefront (bool, optional): Whether to put the request in the beginning or the end of the queue after lock expires
+            lock_secs (int): By how much to prolong the lock, in seconds
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+            forefront=forefront,
+            lockSecs=lock_secs,
+        )
+
+        response = await self.http_client.call(
+            url=self._url(f'requests/{request_id}/lock'),
+            method='PUT',
+            params=request_params,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))
+
+    async def delete_request_lock(self, request_id: str, *, forefront: Optional[bool] = None) -> None:
+        """Delete the lock on a request.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/request-lock/delete-request-lock
+
+        Args:
+            request_id (str): ID of the request to delete the lock
+            forefront (bool, optional): Whether to put the request in the beginning or the end of the queue after the lock is deleted
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+            forefront=forefront,
+        )
+
+        await self.http_client.call(
+            url=self._url(f'requests/{request_id}/lock'),
+            method='DELETE',
+            params=request_params,
+        )
+
+    async def batch_add_requests(self, requests: List[Dict[str, Any]], *, forefront: Optional[bool] = None) -> Dict:
+        """Add requests to the queue.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/batch-request-operations/add-requests
+
+        Args:
+            requests (List[Dict[str, Any]]): List of the requests to add
+            forefront (bool, optional): Whether to add the requests to the head or the end of the queue
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+            forefront=forefront,
+        )
+
+        response = await self.http_client.call(
+            url=self._url('requests/batch'),
+            method='POST',
+            params=request_params,
+            json=requests,
+        )
+        return parse_date_fields(_pluck_data(response.json()))
+
+    async def batch_delete_requests(self, requests: List[Dict[str, Any]]) -> Dict:
+        """Delete given requests from the queue.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/batch-request-operations/delete-requests
+
+        Args:
+            requests (List[Dict[str, Any]]): List of the requests to delete
+        """
+        request_params = self._params(
+            clientKey=self.client_key,
+        )
+
+        response = await self.http_client.call(
+            url=self._url('requests/batch'),
+            method='DELETE',
+            params=request_params,
+            json=requests,
+        )
+        return parse_date_fields(_pluck_data(response.json()))
+
+    async def list_requests(self, *, limit: Optional[int] = None, exclusive_start_id: Optional[str] = None) -> Dict:
+        """List requests in the queue.
+
+        https://docs.apify.com/api/v2#/reference/request-queues/request-collection/list-requests
+
+        Args:
+            limit (int, optional): How many requests to retrieve
+            exclusive_start_id (str, optional): All requests up to this one (including) are skipped from the result
+        """
+        request_params = self._params(limit=limit, exclusive_start_id=exclusive_start_id, clientKey=self.client_key)
+
+        response = await self.http_client.call(
+            url=self._url('requests'),
+            method='GET',
+            params=request_params,
+        )
+
+        return parse_date_fields(_pluck_data(response.json()))